Support undecorated windows in `WindowConfig`
Rationale: Palette style windows are useful for components beyond menus and combo boxes.  This
patch adds support for undecorated windows in WindowConfig.

Use-case: I have a fairly large popup I need to show to select something where a lot of
visual detail needs to be visible to the user; while I have this implemented with `add_overlay()`,
I do not expect that to be usable in a pleasant way if the window size is smaller than what the
popup requires (scrolling would be an ugly solution).