Add the ability to configure mac-os specific properties to WindowConfig
Rationale: Mac OS supports a number of window features that users of Mac OS applications expect, but
which don't make sense as lowest-common-denominator cases.

I am one of the authors of NetBeans, and did *all* of the initial *make it Mac user friendly* circa 2002,
which was extra difficult due to Java AWT's purity-test that if it were a feature supported by Window,
Solaris, Linux *and* Mac OS then it had no place being supported by AWT.  That significantly worsened
the user experience (eventually Apple supported magic "component properties" that made some improvements
possible), and is not a mistake you want to imitate.

Use Case: I have some palette windows that I need more control of the appearance of than is possible
without this.