Breaking out the chainsaw - moved a lot of windowing-specific stuff into the new `windowing` crate.
Essentially, anything possible that touches `winit` belongs there (there are a few stragglers that
need some more thought, like allocating GPU resources, and the field in `WindowHandle` which could
just as easily be obtained from `WindowingSystem`.

Attempting to abstract - and where possible, clean up some of the logic that was scattered about for
calculating locations and managing things.

As a side-effect, `ViewId` now lives in `windowing`, but all of its methods live on a new `ViewIdentifier`
trait in the main crate.  This allows us to have a common implementation of some functionality that
needs to cache based on view ids, but does not need to know much about them.  It does mean clients that
use those methods will need to `use` the new trait.